Created new html pages and edited a file name for img/taco.jpg
added new recipe for a burger and for a taco
added an unordered list for the ingredients needed to create a burger and also added a ordered list on how to prepare the burger.
added an image of a taco and an unordered list for the ingredients needed to create a taco and also added a ordered list on how to prepare the taco.
changed the file name of img.Taco.jpg to img/taco.jpg